+++ src/main/scala/org/broadinstitute/hail/check/Prop.scala
@@ -1,6 +1,7 @@
 package org.broadinstitute.hail.check
 
 import scala.collection.mutable.ArrayBuffer
+import scala.util.Random
 
 abstract class Prop {
   def apply(p: Parameters, name: Option[String] = None): Unit
@@ -12,7 +13,7 @@ abstract class Prop {
       if (!System.getProperty("hail.randomize", "false").toBoolean)
         System.getProperty("hail.seed", "1").toInt
       else
-        Gen.arbInt.sample()
+        Random.nextInt()
     }
 
     println(s"Using a seed of [$seed] for testing.")
@@ -27,7 +28,7 @@ abstract class Prop {
       if (!random && !System.getProperty("hail.randomize", "false").toBoolean)
         seed.getOrElse(System.getProperty("hail.seed", "1").toInt).toLong
       else
-        Gen.arbInt.sample()
+        Random.nextInt()
     }
 
     println(s"Using a seed of [$seed2] for testing.")
@@ -108,7 +109,9 @@ class Properties(val name: String) extends Prop {
 }
 
 object Prop {
-  def check(p: Prop) { p.check() }
+  def check(p: Prop) {
+    p.check()
+  }
 
   def forAll[T1](g1: Gen[T1])(p: (T1) => Boolean): Prop =
     new GenProp1(g1, p)