+++ src/main/scala/org/broadinstitute/hail/methods/ExportTSV.scala
@@ -12,7 +12,7 @@ object ExportTSV {
   def parseColumnsFile(
     ec: EvalContext,
     path: String,
-    hConf: hadoop.conf.Configuration): (Array[String], Array[() => Option[Any]]) = {
+    hConf: hadoop.conf.Configuration): (Array[String], Array[(Type, () => Option[Any])]) = {
     val pairs = readFile(path, hConf) { reader =>
       Source.fromInputStream(reader)
         .getLines()
@@ -27,9 +27,27 @@ object ExportTSV {
 
     val header = pairs.map(_._1)
     val fs = pairs.map { case (_, e) =>
-      Parser.parse(e, ec)._2
+      val (bt, f) = Parser.parse(e, ec)
+      bt match {
+        case t: Type => (t, f)
+        case _ => fatal(s"invalid export type: `$bt'")
+      }
     }
 
     (header, fs)
   }
+
+  def exportTypes(filename: String, hConf: hadoop.conf.Configuration, info: Array[(String, Type)]) {
+    val sb = new StringBuilder
+    writeTextFile(filename, hConf) { out =>
+      info.foreachBetween { case (name, t) =>
+        sb.append(prettyIdentifier(name))
+        sb.append(":")
+        t.pretty(sb, printAttrs = true, compact = true)
+      }(() => sb.append(","))
+
+    out.write(sb.result())
+    }
+  }
+
 }